gr4vy_python/requirements.txt
openapi-client==1.0.0
pem==21.2.0
PyJWT==2.3.0
python-dateutil==2.8.2
six==1.16.0
urllib3==1.26.7
